requirements.txt
python-telegram-bot==20.7
python-vlc==3.0.18121
PyYAML==6.0.1
PySocks==1.7.1
psutil==5.9.6
